Refresh the look and feel of the number pad
In order to match the design of most calculators, the operator keys
were moved to the right-hand side. This will hopefully improve
discoverability of the operators for users. In addition, the look of the
buttons were updated to have a corner radius and the buttons are spaced
out.

The Nib file now uses Stack Views, and as such the deployment target for
the project was set to 9.0.

Some other changes:

* The CalculatorInputView.xib File Owner is now CalculatorInputView and
  it is marked as IB_DESIGNABLE.
* The buttons now have a slight drop shadow
* The size of the text on the buttons was increased 25%

Lastly, in order to hide the drop shadow on touch down and maintain the
corner radius, the `maskToBounds` property is toggled in action methods.