Update tests to Swift 3 and fix configuration error
Some of the tests used syntax used prior to Swift 3 and that was updated
to the current syntax. In addition, the Product Name of the test target
was changed to be CalculatorInput, which had the side effect of making
imports not work correctly. When set back to the previous value of
CalculatorInputTests, the tests are able to find the referenced types
and the tests pass.

The TravisCI configuration is still not correct, so I expect that to
fail.